projects/week10-11-capstone/ingestion/requirements.txt
requests>=2.31.0
faker>=22.0.0
google-cloud-bigquery>=3.14.0
google-cloud-storage>=2.14.0
google-cloud-pubsub>=2.19.0
functions-framework>=3.5.0
pandas>=2.1.0
pyarrow>=14.0.0
great-expectations>=1.0.0
sqlalchemy-bigquery>=1.11.0
